+++ odk2gn/tests/fixtures.py
@@ -437,133 +437,3 @@ def my_config():
 @pytest.fixture(scope="function")
 def attachment():
     return ""
-
-
-@pytest.fixture(scope="function")
-def pf_sub(plant, observers_and_list, nomenclature):
-    pf_sub = [
-        {
-            "__id": uuid.uuid4(),
-            "zp_geom_4326": {
-                "type": "Polygon",
-                "coordinates": [
-                    [
-                        [6.352753, 44.670053, 0, 0],
-                        [6.354138, 44.647097, 0, 0],
-                        [6.370172, 44.669491, 0, 0],
-                        [6.352753, 44.670053, 0, 0],
-                    ]
-                ],
-            },
-            "cd_nom": plant.cd_nom,
-            "date_min": datetime.date.today(),
-            "zp_area": "1758888.79",
-            "observers": "" + str(observers_and_list["user_list"][0].id_role),
-            "aps": [
-                {
-                    "type_geom": "point",
-                    "ap_geom_shape": None,
-                    "ap_geom_point": {
-                        "type": "Point",
-                        "coordinates": [6.355671, 44.655574, 0, 0],
-                    },
-                    "situation": {
-                        "id_nomenclature_incline": nomenclature.id_nomenclature,
-                        "ap_area_shape": None,
-                        "ap_area_point": "0",
-                        "physiognomies": str(nomenclature.id_nomenclature),
-                    },
-                    "habitat": {
-                        "id_nomenclature_habitat": nomenclature.id_nomenclature,
-                        "threat_level": "0",
-                        "favorable_status_percent": 0,
-                        "perturbations": str(nomenclature.id_nomenclature),
-                    },
-                    "id_nomenclature_phenology": nomenclature.id_nomenclature,
-                    "frequency_est": {
-                        "id_nomenclature_frequency_method": nomenclature.id_nomenclature,
-                        "frequency": 0,
-                    },
-                    "count": {
-                        "counting_method": "1",
-                        "total_min": None,
-                        "total_max": None,
-                        "num": 1,
-                    },
-                    "comment": None,
-                },
-                {
-                    "type_geom": "point",
-                    "ap_geom_shape": None,
-                    "ap_geom_point": {
-                        "type": "Point",
-                        "coordinates": [6.355671, 44.655574, 0],
-                    },
-                    "situation": {
-                        "id_nomenclature_incline": nomenclature.id_nomenclature,
-                        "ap_area_shape": None,
-                        "ap_area_point": "0",
-                        "physiognomies": str(nomenclature.id_nomenclature),
-                    },
-                    "habitat": {
-                        "id_nomenclature_habitat": nomenclature.id_nomenclature,
-                        "threat_level": "0",
-                        "favorable_status_percent": 0,
-                        "perturbations": str(nomenclature.id_nomenclature),
-                    },
-                    "id_nomenclature_phenology": nomenclature.id_nomenclature,
-                    "frequency_est": {
-                        "id_nomenclature_frequency_method": nomenclature.id_nomenclature,
-                        "frequency": 0,
-                    },
-                    "count": {
-                        "counting_method": "2",
-                        "total_min": 1,
-                        "total_max": 1,
-                        "num": None,
-                    },
-                    "comment": None,
-                },
-                {
-                    "type_geom": "shape",
-                    "ap_geom_point": None,
-                    "ap_geom_shape": {
-                        "type": "Polygon",
-                        "coordinates": [
-                            [
-                                [6.352753, 44.670053, 0],
-                                [6.354138, 44.647097, 0],
-                                [6.370172, 44.669491, 0],
-                                [6.352753, 44.670053, 0],
-                            ]
-                        ],
-                    },
-                    "situation": {
-                        "id_nomenclature_incline": nomenclature.id_nomenclature,
-                        "ap_area_shape": "1758888.79",
-                        "ap_area_point": None,
-                        "physiognomies": str(nomenclature.id_nomenclature),
-                    },
-                    "habitat": {
-                        "id_nomenclature_habitat": nomenclature.id_nomenclature,
-                        "threat_level": "0",
-                        "favorable_status_percent": 0,
-                        "perturbations": str(nomenclature.id_nomenclature),
-                    },
-                    "id_nomenclature_phenology": nomenclature.id_nomenclature,
-                    "frequency_est": {
-                        "id_nomenclature_frequency_method": "mille",
-                        "frequency": 0,
-                    },
-                    "count": {
-                        "counting_method": "0",
-                        "total_min": None,
-                        "total_max": None,
-                        "num": None,
-                    },
-                    "comment": None,
-                },
-            ],
-        },
-    ]
-    return pf_sub
